refactor: specialValue storage to long
Former-commit-id: 7e7b67c62751c4d1e8af1e51d0f912a9aecaf4c8

+++ Allay-API/src/main/java/cn/allay/api/block/property/type/BaseBlockPropertyType.java
@@ -13,9 +13,9 @@ public abstract sealed class BaseBlockPropertyType<DATATYPE> implements BlockPro
     protected final String name;
     protected final List<DATATYPE> validValues;
     protected final DATATYPE defaultValue;
-    protected final int bitSize;
+    protected final byte bitSize;
 
-    protected BaseBlockPropertyType(String name, List<DATATYPE> validValues, DATATYPE defaultValue, int bitSize) {
+    protected BaseBlockPropertyType(String name, List<DATATYPE> validValues, DATATYPE defaultValue, byte bitSize) {
         Objects.requireNonNull(defaultValue);
         this.name = name;
         this.validValues = validValues;
@@ -39,7 +39,7 @@ public abstract sealed class BaseBlockPropertyType<DATATYPE> implements BlockPro
     }
 
     @Override
-    public int getBitSize() {
+    public byte getBitSize() {
         return bitSize;
     }
 }
