+++ Allay-API/src/main/java/cn/allay/api/block/property/type/BlockPropertyType.java
@@ -32,7 +32,7 @@ public sealed interface BlockPropertyType<DATATYPE> permits BaseBlockPropertyTyp
 
     BlockPropertyValue<DATATYPE, ? extends BlockPropertyType<DATATYPE>, ?> tryCreateValue(Object value);
 
-    int getBitSize();
+    byte getBitSize();
 
     default BlockPropertyValue<DATATYPE, ? extends BlockPropertyType<DATATYPE>, ?> createDefaultValue() {
         return createValue(getDefaultValue());
